Fix type mismatches

+++ libatalk/cnid/mtab/cnid_mtab.c
@@ -1,6 +1,6 @@
 
 /*
- * $Id: cnid_mtab.c,v 1.2 2005-04-28 20:50:01 bfernhomberg Exp $
+ * $Id: cnid_mtab.c,v 1.3 2009-11-24 08:59:25 franklahm Exp $
  *
  * Copyright (c) 1999. Adrian Sun (asun@zoology.washington.edu)
  * All Rights Reserved. See COPYRIGHT.
@@ -31,7 +31,7 @@
 
 /* ------------------------ */
 cnid_t cnid_mtab_add(struct _cnid_db *cdb, const struct stat *st,
-                     const cnid_t did, char *name, const int len, cnid_t hint)
+                     const cnid_t did, char *name, const size_t len, cnid_t hint)
 {
     struct stat lst;
     const struct stat *lstp;
@@ -59,13 +59,13 @@ int cnid_mtab_delete(struct _cnid_db *cdb, const cnid_t id)
 
 
 /* Return CNID for a given did/name. */
-cnid_t cnid_mtab_get(struct _cnid_db *cdb, const cnid_t did, char *name, const int len)
+cnid_t cnid_mtab_get(struct _cnid_db *cdb, const cnid_t did, char *name, const size_t len)
 {
     return CNID_INVALID;
 }
 
 
-cnid_t cnid_mtab_lookup(struct _cnid_db *cdb, const struct stat *st, const cnid_t did, char *name, const int len)
+cnid_t cnid_mtab_lookup(struct _cnid_db *cdb, const struct stat *st, const cnid_t did, char *name, const size_t len)
 {
     return CNID_INVALID;
 }
@@ -121,14 +121,14 @@ struct _cnid_module cnid_mtab_module = {
 
 
 /* Return the did/name pair corresponding to a CNID. */
-char *cnid_mtab_resolve(struct _cnid_db *cdb, cnid_t * id, void *buffer, u_int32_t len)
+char *cnid_mtab_resolve(struct _cnid_db *cdb, cnid_t * id, void *buffer, size_t len)
 {
     return NULL;
 }
 
 
 int cnid_mtab_update(struct _cnid_db *cdb, const cnid_t id, const struct stat *st,
-                     const cnid_t did, char *name, const int len
+                     const cnid_t did, char *name, const size_t len
                      /*, const char *info, const int infolen */ )
 {
     return 0;
